LPD-58785 Move naming to Database Migration Importer.
