+++ modules/util/portal-tools-db-migration-importer-test/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Portal Tools DB Migration Importer Test
+Bundle-SymbolicName: com.liferay.portal.tools.db.migration.importer.test
+Bundle-Version: 1.0.0
+-removeheaders: Require-Capability
