+++ modules/util/portal-tools-db-migration-importer-test/src/testFunctional/tests/DBMigrationImport.testcase
@@ -1,0 +1,134 @@
+@component-name = "portal-db-infrastructure"
+definition {
+
+	property app.server.types = "tomcat";
+	property custom.properties = "LPD-23840=true";
+	property database.types = "db2,mariadb,mysql,oracle,sqlserver";
+	property portal.release = "true";
+	property portal.upstream = "true";
+	property test.assert.warning.exceptions = "true";
+	property test.example = "true";
+	property test.liferay.virtual.instance = "true";
+	property test.run.environment = "EE";
+	property test.run.type = "single";
+	property testray.main.component.name = "Database Upgrade Framework";
+
+	setUp {
+		TestCase.setUpPortalInstance();
+
+		User.firstLoginPG();
+	}
+
+	tearDown {
+		PortalInstances.tearDownCP();
+
+		DBMigrationExportImport.tearDown();
+	}
+
+	@priority = 5
+	test CanImportDatabaseToPostgreSQL {
+		task ("Export origin database schema definition") {
+			DBMigrationExportImport.dbSchemaDefinitionExport();
+		}
+
+		task ("Sign out and stop server") {
+			User.logoutPG();
+
+			Portlet.shutdownServer();
+		}
+
+		task ("Start PostgreSQL database and execute database import to PostgreSQL") {
+			DBMigrationExportImport.startPostgreSQLDB();
+
+			DBMigrationExportImport.executeDatabaseImport();
+		}
+
+		task ("Start server pointing to PostgreSQL database") {
+			DBMigrationExportImport.switchToPostgreSQLDB();
+
+			Portlet.startServer(deleteLiferayHome = "false");
+		}
+
+		task ("Ensure that portal main functionalities work correctly") {
+			SignIn.signInTestSetup();
+
+			Smoke.runSmoke();
+
+			SignOut.signOut();
+		}
+	}
+
+	@priority = 5
+	test CanImportPartitionedDatabaseToPostgreSQL {
+		property custom.properties = "database.partition.enabled=true{line.separator}feature.flag.LPD-23840=true${line.separator}virtual.hosts.valid.hosts=localhost,127.0.0.1,www.able.com";
+		property database.types = "mysql";
+		property test.liferay.virtual.instance = "false";
+
+		task ("Add www.able.com with an asset") {
+			HeadlessPortalInstanceAPI.addPortalInstance(
+				domain = "www.able.com",
+				portalInstanceId = "www.able.com",
+				virtualHost = "www.able.com");
+
+			User.firstLoginUI(
+				password = PropsUtil.get("default.admin.password"),
+				specificURL = "http://www.able.com:8080",
+				userEmailAddress = "test@www.able.com");
+
+			ProductMenu.gotoPortlet(
+				category = "Content & Data",
+				panel = "Site Administration",
+				portlet = "Documents and Media");
+
+			DMDocument.addCP(
+				dmDocumentDescription = "DM Document Description",
+				dmDocumentFile = "Document_1.doc",
+				dmDocumentTitle = "DM Document Title");
+
+			DMNavigator.gotoDocumentCP(dmDocumentTitle = "DM Document Title");
+
+			DMDocument.viewCP(dmDocumentTitle = "DM Document Title");
+
+			SignOut.signOutOfSpecificURL(url = "http://www.able.com:8080");
+		}
+
+		task ("Export origin database schema definition") {
+			DBMigrationExportImport.dbSchemaDefinitionExport();
+		}
+
+		task ("Sign out and stop server") {
+			User.logoutPG();
+
+			Portlet.shutdownServer();
+		}
+
+		task ("Start PostgreSQL database and execute database import to PostgreSQL") {
+			DBMigrationExportImport.startPostgreSQLDB();
+
+			DBMigrationExportImport.executeDatabaseImport();
+		}
+
+		task ("Start server pointing to PostgreSQL database") {
+			DBMigrationExportImport.switchToPostgreSQLDB();
+
+			Portlet.startServer(deleteLiferayHome = "false");
+		}
+
+		task ("Ensure that www.able.com is accessible and its assets visible") {
+			User.firstLoginUI(
+				password = PropsUtil.get("default.admin.password"),
+				specificURL = "http://www.able.com:8080",
+				userEmailAddress = "test@www.able.com");
+
+			ProductMenu.gotoPortlet(
+				category = "Content & Data",
+				panel = "Site Administration",
+				portlet = "Documents and Media");
+
+			DMNavigator.gotoDocumentCP(dmDocumentTitle = "DM Document Title");
+
+			DMDocument.viewCP(dmDocumentTitle = "DM Document Title");
+		}
+	}
+
+}
