--- modules/util/portal-tools-db-migration-schema-importer-test/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: Liferay Portal Tools DB Migration Schema Importer Test
-Bundle-SymbolicName: com.liferay.portal.tools.db.migration.schema.importer.test
-Bundle-Version: 1.0.0
--removeheaders: Require-Capability
