--- modules/util/portal-tools-db-migration-schema-importer/source-formatter-suppressions.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0"?>
-
-<suppressions>
-	<source-check>
-		<suppress checks="IllegalImportsCheck" files="src/main/java/com/liferay/portal/tools/db/migration/schema/importer/jdbc/AutoBatchPreparedStatementUtil\.java" />
-	</source-check>
-</suppressions>
